+++ docs/src/history.md
@@ -4,6 +4,7 @@
 ## v0.2.1 (current development version)
 
 * ADDED: `@critical` macro to make test assertions terminate the testcase on failure.
+* ADDED: for `verbosity=1`, display the results returned by `@test_result` separately.
 * FIXED: `@produce` hanging when passed a non-string label.
 * FIXED: incorrect indentation with `verbosity=1` when a group has some testcases after nested groups.